requirements.txt
asgiref==3.2.3
autopep8==1.4.4
beautifulsoup4==4.8.2
Django==3.0.2
django-admin-rangefilter==0.5.3
django-cors-headers==3.2.1
django-htmlmin==0.11.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
gunicorn==20.0.4
html5lib==1.0.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pytz==2019.3
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
webencodings==0.5.1
